Correcciones en el inicio de sesión
